+++ app/Models/DatabaseUpgrades.php
@@ -131,6 +131,21 @@ class DatabaseUpgrades extends Models {
     $sql_data[] = "
       ALTER TABLE `".PREFIX."users` ADD `location` VARCHAR(255) NULL DEFAULT NULL AFTER `website`;
     ";
+    $sql_data[] = "
+      CREATE TABLE IF NOT EXISTS `".PREFIX."comments` (
+        `id` int(10) NOT NULL AUTO_INCREMENT,
+        `com_id` int(10) DEFAULT NULL,
+        `com_sec_id` int(10) DEFAULT NULL,
+        `com_location` varchar(255) DEFAULT NULL,
+        `com_user_ip` varchar(50) DEFAULT NULL,
+        `com_server` varchar(255) DEFAULT NULL,
+        `com_uri` varchar(255) DEFAULT NULL,
+        `com_owner_userid` int(10) DEFAULT NULL,
+        `com_content` text DEFAULT NULL,
+        `timestamp` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
+        PRIMARY KEY (`id`)
+      ) ENGINE=InnoDB  DEFAULT CHARSET=latin1 AUTO_INCREMENT=1 ;
+    ";
 
     foreach ($sql_data as $query) {
       if(!$this->db->upgrade($query)){ $error[] = true; }